requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
fastapi==0.115.4
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
-e git+https://github.com/cvg/GeoCalib@a7ef2cfcd06ce2b1628ee2c64f54bdaf3fff81cd#egg=geocalib
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.5
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.1
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pillow==11.0.0
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-multipart==0.0.17
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
sympy==1.13.1
torch==2.5.0
torchvision==0.20.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
Viz2D==0.0.1
